# Limit amount of channel buffers