@paulmach 2021-02-02
This is a good find. Definitely N^2 is not good because every N is 8000 osm elements in memory.